% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_2d.R
\name{add_slice_2d}
\alias{add_slice_2d}
\title{2D Slice to R Plot}
\usage{
add_slice_2d(
  model,
  x_axis = NA,
  n = 100,
  caption = FALSE,
  back_transform = TRUE,
  ...,
  xaxis = NULL
)
}
\arguments{
\item{model}{A linear model}

\item{x_axis}{A non-required specification of the x-axis variable; default is first x var in lm}

\item{n}{Number of displayed points; default is 100}

\item{caption}{Boolean to include caption with values of predictor variables not shown in the chart, FALSE by default}

\item{back_transform}{Either 1) A function to back transform the predicted y values and interval bounds (e.g., \code{exp} or \verb{\\(x) exp(x)} for log-transformed models), or 2) a boolean stating whether it should apply a transformation or not (\code{FALSE} turns off back_transform). In most cases, this does not need to be specified as the function infers the transformation needed reliably.}

\item{...}{Remaining vars (color, linetype, etc.)}

\item{xaxis}{Deprecated; use \code{x_axis} instead.}
}
\value{
A line added to a preexisting R plot displaying a 2D graphable slice of an HD model
}
\description{
Take a slice from a linear model and add the displayed graph to a preexisting R plot.
This function is similar to slice_2d() but adds a line to an existing plot rather
than creating a new one. Useful for overlaying multiple model slices or adding
model predictions to scatter plots.
}
\examples{
\dontrun{
model <- lm(mpg ~ disp + hp, data = mtcars)
plot(mpg ~ disp, data = mtcars)
add_slice_2d(model,x_axis='disp', hp=100)
}

}
