% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inverse_function.R
\name{get_inverse_function}
\alias{get_inverse_function}
\title{Generate an Inverse Function from a Formula LHS}
\usage{
get_inverse_function(f)
}
\arguments{
\item{f}{A formula (e.g., log(y) ~ x) or an expression/call.}
}
\value{
A function accepting one argument (x) that returns the back-transformed value.
}
\description{
This function analyzes the Left-Hand Side (LHS) of a formula and constructs
a function that performs the mathematical inverse.
}
