% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_2d.R
\name{add_slice_2d}
\alias{add_slice_2d}
\title{2D Slice to R Plot}
\usage{
add_slice_2d(model, xaxis = NA, n = 100, ...)
}
\arguments{
\item{model}{A linear model}

\item{xaxis}{A non-required specification of the xaxis variable; default is first x var in lm}

\item{n}{Number of displayed points; default is 100}

\item{...}{}
}
\value{
A line added to a preexisting R plot displaying a 2D graphable slice of an HD model
}
\description{
Take a slice from a linear model and add the displayed graph to a preexisting R plot
}
\examples{
kids <- lm(width ~ length + birthmonth, data = KidsFeet)
plot(width ~ length, data = KidsFeet)
add_slice_2d(kids)

}
