% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_2d.R
\name{add_slice_2d}
\alias{add_slice_2d}
\title{2D Slice to R Plot}
\usage{
add_slice_2d(model, xaxis = NA, n = 100, ...)
}
\arguments{
\item{model}{A linear model}

\item{xaxis}{A non-required specification of the xaxis variable; default is first x var in lm}

\item{n}{Number of displayed points; default is 100}

\item{...}{Remaining vars (color, linetype, etc.)}
}
\value{
A line added to a preexisting R plot displaying a 2D graphable slice of an HD model
}
\description{
Take a slice from a linear model and add the displayed graph to a preexisting R plot
}
\examples{
\dontrun{
model <- lm(mpg ~ disp + hp, data = mtcars)
plot(mpg ~ disp, data = mtcars)
add_slice_2d(model,xaxis='disp', hp=100)
}

}
