% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Diagnostic plots for linear models}
\usage{
diagnose(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\value{
None. The function is called for its side effect of producing plots.
}
\description{
Generates a set of three diagnostic plots for a linear model object:
Residuals vs Fitted, Normal Q-Q, and a plot of the residuals.
}
\examples{
lm_model <- lm(mpg ~ wt, data = mtcars)
diagnose(lm_model)
}
