% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawit.R
\name{drawit}
\alias{drawit}
\title{Draw It function}
\usage{
drawit(model, xaxis, ..., col = "black", lty = 1)
}
\arguments{
\item{model}{A linear model using lm()}

\item{xaxis}{The xaxis of the plot and primary explanatory variable}

\item{...}{Remaing variables; will throw warnings if not properly declared}

\item{col}{Color of the line}

\item{lty}{Line type}
}
\value{
A slice of a linear model in a base R plot
}
\description{
A function to draw singular slices of a model in base R
}
\examples{
model <- lm(width ~ length + sex + birthmonth, KidsFeet)
plot(width ~ length, KidsFeet)
drawit(model = mode, xaxis = 'length',sex = 'B',birthmonth = 6, col='blue',lty=1)

}
