% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawit.R
\name{drawit}
\alias{drawit}
\title{Draw It function}
\usage{
drawit(model, xaxis, ..., col = "black", lty = 1)
}
\arguments{
\item{model}{A linear model using lm()}

\item{xaxis}{The xaxis of the plot and primary explanatory variable}

\item{...}{Remaining variables; will throw warnings if not properly declared}

\item{col}{Color of the line}

\item{lty}{Line type}
}
\value{
A slice of a linear model in a base R plot
}
\description{
A function to draw singular slices of a model in base R. This function takes a more
strict approach by building the prediction equation from model coefficients
and requiring users to specify values for all predictor variables. Its original intent
is to be most useful for regression battleship, and as a learning tool. By requiring
entry of every value, it requires the student to think about what they are plotting.
}
\examples{
\dontrun{
model <- lm(mpg ~ wt + am + qsec, mtcars)
plot(mpg ~ wt, mtcars)
drawit(model = model, xaxis = 'wt',am = 1,qsec = 17, col='blue',lty=1)
}

}
