% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_slice.R
\name{geom_slice}
\alias{geom_slice}
\title{Geom_slice}
\usage{
geom_slice(model, n = 100, inherit.aes = TRUE, predict_vars = list(), ...)
}
\arguments{
\item{model}{A linear model provided by 'lm()'}

\item{n}{The number of predicted points across the graph - default 100}

\item{inherit.aes}{Inherits the aes provided by 'ggplot' object}

\item{predict_vars}{A list of specified variables and values that specify the slice - not required}

\item{...}{Other variables (color, alpha, etc.)}
}
\value{
A ggplot layer that displays the slice of the high dimensional model
}
\description{
A 'ggplot' layer that takes a high dimensional model and displays the predicted
model across each facet of a graph
}
\examples{
\dontrun{
# Example: Using geom_slice without specifying 'predict_vars'
model <- lm(mpg ~ disp + hp, data = mtcars)
ggplot(mtcars, aes(x=disp,y=mpg))+
geom_point()+
facet_wrap(~hp)+
geom_slice(model=model)

# Example: Using geom_slice with 'predict_vars'
model <- lm(mpg ~ disp + hp + gear, data = mtcars)
ggplot(mtcars, aes(x=disp, y=mpg))+
geom_point()+
facet_wrap(~hp)+
geom_slice(model=model, predict_vars=list(gear=4))
}

}
