% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_slice.R
\docType{data}
\name{GeomSlice}
\alias{GeomSlice}
\title{GeomSlice}
\format{
An object of class \code{GeomSlice} (inherits from \code{GeomLine}, \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomSlice
}
\description{
A custom 'ggproto' object that provides the graphing aesthetics
for the 'StatSlice' object. This defines how the slice lines should look.
It is essentially GeomLine with modified aesthetics
}
\keyword{datasets}
