% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_equation.R
\name{lm_equation}
\alias{lm_equation}
\title{Write out the equation of a linear model}
\usage{
lm_equation(model)
}
\arguments{
\item{model}{A linear model}
}
\value{
The equation of an indicated linear model with coefficients and variable names
}
\description{
Write out the equation of a linear model
}
\examples{
\dontrun{
model <- lm(mpg ~ disp + hp, data = mtcars)
lm_equation(model)
}

}
