% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d.R
\name{scatter_3d}
\alias{scatter_3d}
\title{Plot a 3D-grapheable linear model}
\usage{
scatter_3d(model, n = 100, colors = c("blue", "yellow"))
}
\arguments{
\item{model}{A saved linear model
Ex. df <- lm(y ~ x + x2, data)}

\item{n}{Number of evaluations}

\item{colors}{Colors used for the gradient scale
colors = c(color1, color2, etc.)}
}
\value{
A 3D graph in the viewer section
}
\description{
Plot a 3D-grapheable linear model
}
\examples{
model <- lm(width ~ length + sex, data = KidsFeet)
scatter_3d(model, colors=c('blue','yellow')
}
