% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_2d.R
\name{slice_2d}
\alias{slice_2d}
\title{2D Slicer}
\usage{
slice_2d(model, xaxis = NA, n = 100, ...)
}
\arguments{
\item{model}{A linear model}

\item{xaxis}{A non-required specification of the xaxis variable; default is first x var in lm}

\item{n}{Number of displayed points; default is 100}

\item{...}{Remaining vars (color, linetype, etc.)}
}
\value{
A 2D graph of a sliced model
}
\description{
Take a linear model and display a 2D slice in base R
}
\examples{
\dontrun{
model <- lm(mpg ~ disp, data = mtcars)
slice_2d(model, xaxis = 'disp')
}

}
